+++ packages/icons/src/json-imports/Icons.ts
@@ -4,9 +4,9 @@ const loadIconsBundle = async (collection: string): Promise<CollectionData> => {
     let iconData: CollectionData;
 
 	if (collection === "SAP-icons-v5") {
-		iconData = (await import("../generated/assets/v5/SAP-icons.json")).default;
+		iconData = (await import(/* webpackChunkName: "ui5-webcomponents-sap-icons-v5" */ "../generated/assets/v5/SAP-icons.json")).default;
 	} else {
-		iconData = (await import("../generated/assets/v4/SAP-icons.json")).default;
+		iconData = (await import(/* webpackChunkName: "ui5-webcomponents-sap-icons-v4" */ "../generated/assets/v4/SAP-icons.json")).default;
 	}
 
     if (typeof iconData === "string" && (iconData as string).endsWith(".json")) {